# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API Documentation for the InterMine Registry RESTful API.",
    "version": "1.0.0",
    "title": "Intermine Registry",
    "license": {
      "name": "GNU General Public License v3.0",
      "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
    }
  },
  "host": "registry.intermine.org",
  "basePath": "/service",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/instances/{instancesId}": {
      "put": {
        "tags": [
          "instances"
        ],
        "summary": "Update an instance on the registry",
        "description": "Update the instance entry given in the URL from the registry.",
        "operationId": "updateInstance",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "instancesId",
            "in": "path",
            "description": "ID of the instance that needs to be updated",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Information needed to update an instance on the Registry. Not all the fields are needed.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                    "type": "string",
                    "example": "FlyMine"
                },
                "neighbours": {
                    "items": {
                        "type": "string"
                    },
                    "example": ["MODs"],
                    "type": "array"
                },
                "organisms": {
                    "items": {
                        "type": "string"
                    },
                    "example": ["Drosophila"],
                    "type": "array"
                },
                "url": {
                    "type": "string",
                    "example": "http://www.flymine.org/query"
                },
                "description": {
                    "type": "string",
                    "example": "FlyMine is an integrated database of genomic, expression and protein data for Drosophila, Anopheles and C. elegans"
                },
                "location": {
                    "properties": {
                        "latitude": {
                            "type": "string",
                            "example": "52.2003399"
                        },
                        "longitude": {
                            "type": "string",
                            "example": "0.120109"
                        }
                    },
                    "type": "object"
                },
                "twitter": {
                    "type": "string",
                    "example": "@intermineorg"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Instance Successfully Updated on the Registry"
          },
          "400": {
            "description": "Malformed JSON. Invalid or Missing Data. Check for invalid URL."
          },
          "404": {
            "description": "Instance Not Found"
          },
          "406": {
            "description": "Bad Request"
          }
        },
        "security": [
        ]
      },
      "delete": {
        "tags": [
          "instances"
        ],
        "summary": "Deletes an instance",
        "description": "Delete the instance entry given in the URL from the registry.",
        "operationId": "deleteinstances",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "instancesId",
            "in": "path",
            "description": "Id of the instance to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Instance Successfully Deleted"
          },
          "404": {
            "description": "Instance Not Found"
          }
        },
        "security": [
        ]
      }
    },
...
  },
  "securityDefinitions": {
  },
...
}